move pose_calculator.hpp to include folder
